lms_backend/requirements.txt
# FastAPI and ASGI Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
psycopg2==2.9.10
psycopg==3.2.9
requests==2.32.4

# Database (PostgreSQL support)
sqlalchemy>=2.0.0
alembic>=1.12.0
# PostgreSQL drivers (commented out for Python 3.13 compatibility)
# asyncpg>=0.29.0
# psycopg2-binary>=2.9.9
# Using psycopg3 for Python 3.13 compatibility
psycopg[binary]>=3.1.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Caching and Sessions (commented out for now)
# redis>=5.0.0
# aioredis>=2.0.0

# HTTP Client
httpx>=0.25.0

# File Upload and Storage (commented out for now)
# boto3>=1.34.0
# python-magic>=0.4.27

# Background Tasks (commented out for now)
# celery>=5.3.0

# Email (commented out for now)
# fastapi-mail>=1.4.0

# WebSockets (commented out for now)
# websockets>=12.0

# Testing (commented out for now)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.1.0

# Development Tools (commented out for now)
# black>=23.11.0
# flake8>=6.1.0
# mypy>=1.7.0
# isort>=5.12.0

# Environment and Configuration
python-dotenv>=1.0.0

# Monitoring and Logging
loguru>=0.7.0

# Date and Time
python-dateutil>=2.8.0
pytz>=2023.0

# Utilities
python-slugify>=8.0.0
email-validator>=2.2.0
asyncpg